Bug 1081401: Remove unnecessary autoboxing. r=rnewman
UltraBlame original commit: adc99de50f0dc1461b1f80cc1b55be9df3123411

+++ mobile/android/base/db/BrowserProvider.java
@@ -1065,7 +1065,7 @@ public class BrowserProvider extends SharedBrowserDatabaseProvider {
                     Long additional = values.getAsLong(History.VISITS);
 
                     
-                    values.put(History.VISITS, existing + ((additional != null) ? additional.longValue() : 1));
+                    values.put(History.VISITS, existing + ((additional != null) ? additional : 1));
                 }
 
                 updated += db.update(TABLE_HISTORY, values, "_id = ?",
